requirements.txt
Flask>=2.2,<3.0
Flask-SQLAlchemy>=3.0,<4.0
Flask-Migrate>=4.0,<5.0
Flask-Caching>=2.0.0,<3.0
click>=8.0
itsdangerous>=2.1
Jinja2>=3.0
MarkupSafe>=2.1
python-dateutil>=2.8
SQLAlchemy>=2.0
Werkzeug>=2.3
python-dotenv==1.0.1
llama-cpp-python==0.2.55
